Add api/shops_controller and refactor
