+++ app/assets/javascripts/darkswarm/controllers/producer_node_controller.js.coffee
@@ -24,7 +24,7 @@ Darkswarm.controller "ProducerNodeCtrl", ($scope, HashNavigation, $anchorScroll,
     $scope.shopfront_loading = true
     $scope.toggle_tab(event)
 
-    $http.get("/api/enterprises/" + $scope.producer.id + "/shopfront")
+    $http.get("/api/shops/" + $scope.producer.id)
       .success (data) ->
         $scope.shopfront_loading = false
         $scope.producer = data
