+++ app/assets/javascripts/darkswarm/services/enterprise_modal.js.coffee
@@ -5,7 +5,7 @@ Darkswarm.factory "EnterpriseModal", ($modal, $rootScope, $http)->
       scope = $rootScope.$new(true) # Spawn an isolate to contain the enterprise
       scope.embedded_layout = window.location.search.indexOf("embedded_shopfront=true") != -1
 
-      $http.get("/api/enterprises/" + enterprise.id + "/shopfront").success (data) ->
+      $http.get("/api/shops/" + enterprise.id).success (data) ->
         scope.enterprise = data
         $modal.open(templateUrl: "enterprise_modal.html", scope: scope)
       .error (data) ->
