+++ app/assets/javascripts/darkswarm/services/enterprise_resource.js.coffee
@@ -5,8 +5,4 @@ Darkswarm.factory 'EnterpriseResource', ($resource) ->
       url: '/enterprises/:id/relatives.json'
       isArray: true
       cache: true
-    'closed_shops':
-      method: 'GET'
-      isArray: true
-      url: '/api/enterprises/closed_shops.json'
   })
