+++ app/assets/javascripts/darkswarm/services/shops_resource.js.coffee
@@ -1,0 +1,7 @@
+Darkswarm.factory 'ShopsResource', ($resource) ->
+  $resource('/api/shops/:id.json', {}, {
+    'closed_shops':
+      method: 'GET'
+      isArray: true
+      url: '/api/shops/closed_shops.json'
+  })
