+++ config/routes/api.rb
@@ -33,11 +33,9 @@ Openfoodnetwork::Application.routes.draw do
 
       resource :logo, only: [:destroy]
       resource :promo_image, only: [:destroy]
+    end
 
-      member do
-        get :shopfront
-      end
-
+    resources :shops, only: [:show] do
       collection do
         get :closed_shops
       end
