+++ spec/controllers/api/enterprises_controller_spec.rb
@@ -81,41 +81,5 @@ module Api
         end
       end
     end
-
-    context "as a non-authenticated user" do
-      let!(:hub) {
-        create(:distributor_enterprise, with_payment_and_shipping: true, name: 'Shopfront Test Hub')
-      }
-      let!(:producer) { create(:supplier_enterprise, name: 'Shopfront Test Producer') }
-      let!(:category) { create(:taxon, name: 'Fruit') }
-      let!(:product) { create(:product, supplier: producer, primary_taxon: category ) }
-      let!(:relationship) { create(:enterprise_relationship, parent: hub, child: producer) }
-      let!(:closed_hub1) { create(:distributor_enterprise) }
-      let!(:closed_hub2) { create(:distributor_enterprise) }
-
-      before do
-        allow(controller).to receive(:spree_current_user) { nil }
-      end
-
-      describe "#shopfront" do
-        it "returns shopfront data for an enterprise" do
-          spree_get :shopfront, id: producer.id, format: :json
-
-          expect(json_response['name']).to eq 'Shopfront Test Producer'
-          expect(json_response['hubs'][0]['name']).to eq 'Shopfront Test Hub'
-          expect(json_response['supplied_taxons'][0]['name']).to eq 'Fruit'
-        end
-      end
-
-      describe "#closed_shops" do
-        it "returns data for all closed shops" do
-          spree_get :closed_shops, nil, format: :json
-
-          expect(json_response).not_to match hub.name
-          expect(json_response[0]['id']).to eq closed_hub1.id
-          expect(json_response[1]['id']).to eq closed_hub2.id
-        end
-      end
-    end
   end
 end
